Improved DataFormNumberWidget

+++ src/Widgets/DataFormNumberWidget/DataFormNumberWidgetClass.ts
@@ -4,8 +4,8 @@ export const DataFormNumberWidget = provideWidgetClass('DataFormNumberWidget', {
   attributes: {
     attributeName: 'string',
     placeholder: 'string',
-    minValue: 'integer',
-    maxValue: 'integer',
+    minValue: 'float',
+    maxValue: 'float',
     stepValue: 'float',
     helpText: 'html',
     label: 'string',
